Rebased and applied review comments
